+++ extensions/immich/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Immich Changelog
 
+## [Copy and Download assets] - {PR_MERGE_DATE}
+
+- Add Copy and Download actions for assets in the Asset View
+
 ## [Initial Version] - 2025-11-14
 
 - Explore
-- Library 
+- Library
 - API Keys